Wrap search param usage in suspense
