+++ app/page.tsx
@@ -1,45 +1,10 @@
-"use client";
-import { useEffect, useRef } from "react";
-import { useSearchParams } from "next/navigation";
-import ChatPane from "@/components/panels/ChatPane";
-import MedicalProfile from "@/components/panels/MedicalProfile";
-import Timeline from "@/components/panels/Timeline";
-import AlertsPane from "@/components/panels/AlertsPane";
-import SettingsPane from "@/components/panels/SettingsPane";
+import { Suspense } from "react";
+import MainPageContent from "@/components/MainPageContent";
 
 export default function Page() {
-  const searchParams = useSearchParams();
-  const panel = (searchParams.get("panel") ?? "chat").toLowerCase();
-  const threadId = searchParams.get("threadId") ?? undefined;
-  const chatInputRef = useRef<HTMLInputElement>(null);
-
-  useEffect(() => {
-    const handler = () => chatInputRef.current?.focus();
-    window.addEventListener("focus-chat-input", handler);
-    return () => window.removeEventListener("focus-chat-input", handler);
-  }, []);
-
   return (
-    <>
-      <section className={panel === "chat" ? "block h-full" : "hidden"}>
-        <ChatPane inputRef={chatInputRef} />
-      </section>
-
-      <section className={panel === "profile" ? "block" : "hidden"}>
-        <MedicalProfile />
-      </section>
-
-      <section className={panel === "timeline" ? "block" : "hidden"}>
-        <Timeline threadId={threadId} />
-      </section>
-
-      <section className={panel === "alerts" ? "block" : "hidden"}>
-        <AlertsPane />
-      </section>
-
-      <section className={panel === "settings" ? "block" : "hidden"}>
-        <SettingsPane />
-      </section>
-    </>
+    <Suspense fallback={<div />}>
+      <MainPageContent />
+    </Suspense>
   );
 }
